refactor(permissions-&-access-rights): apply some changes requested
